Snapshot worldgen and HUD updates

+++ src/Game/Core/FactionDirector.Targets.cs
@@ -4,10 +4,16 @@ public sealed partial class FactionDirector
 {
     private HexCoord? ChooseStackTarget(GameState state, StackState stack, string actionId, Random random)
     {
+        // Targets are selected from the stack's current movement range. The AI
+        // never plans multi-turn paths yet; it chooses the best reachable hex
+        // for this one turn.
         var range = GameRules.MovementRange(state, stack.Coord, stack.MovementLeft);
 
         if (actionId is "attack_enemy" or "defend_city")
         {
+            // Both attack and defense currently move toward enemy stacks. This
+            // makes "defend city" a reactive military posture until stronger
+            // garrison/zone rules exist.
             var enemies = state.Stacks.Values
                 .Where(s => s.FactionId != stack.FactionId)
                 .OrderBy(s => s.Coord.DistanceTo(stack.Coord))
@@ -21,6 +27,9 @@ public sealed partial class FactionDirector
 
         if (actionId == "claim_resource")
         {
+            // Resource claiming moves toward the nearest visible resource with a
+            // small random tie-breaker so AI factions do not always pick the same
+            // exact target in similar positions.
             var resource = state.Map.Tiles
                 .Where(t => t.ResourceId is not null)
                 .OrderBy(t => t.Coord.DistanceTo(stack.Coord) + random.Next(0, 3))
@@ -34,31 +43,40 @@ public sealed partial class FactionDirector
             .OrderBy(c => c.Coord.DistanceTo(stack.Coord))
             .FirstOrDefault();
 
+        // Upgrade-focused stacks drift toward their own city; otherwise they
+        // wander within range to keep the board moving.
         return actionId == "upgrade_city" && ownedCity is not null
             ? ClosestReachable(range, ownedCity.Coord)
             : RandomReachable(range, random);
     }
 
-    private HexCoord? ChooseScoutTarget(GameState state, HexCoord origin, int movement, Random random)
+    private HexCoord? ChooseScoutTarget(GameState state, HexCoord origin, double movement, Random random)
     {
+        // Scouts are intentionally simple: pick any reachable non-origin tile.
+        // Fog of war can later replace this with unknown-tile exploration.
         var range = GameRules.MovementRange(state, origin, movement);
         return RandomReachable(range, random);
     }
 
-    private static HexCoord? RandomReachable(Dictionary<HexCoord, int> range, Random random)
+    private static HexCoord? RandomReachable(Dictionary<HexCoord, double> range, Random random)
     {
+        // Exclude the origin, which is always present with cost 0.
         var options = range.Keys.Where(c => range[c] > 0).ToList();
         return options.Count == 0 ? null : options[random.Next(options.Count)];
     }
 
-    private static HexCoord? ClosestReachable(Dictionary<HexCoord, int> range, HexCoord target)
+    private static HexCoord? ClosestReachable(Dictionary<HexCoord, double> range, HexCoord target)
     {
+        // The reachable hex closest to the target is the one-turn approximation
+        // of pathfinding toward that target.
         var options = range.Keys.Where(c => range[c] > 0).ToList();
         return options.Count == 0 ? null : options.OrderBy(c => c.DistanceTo(target)).First();
     }
 
     private void TryUpgradeCity(GameState state, string factionId, Random random)
     {
+        // The director's chosen action can bias movement toward a city, but the
+        // actual upgrade still has a chance gate to keep AI turns varied.
         var city = state.Cities.Values.FirstOrDefault(c => c.FactionId == factionId);
         if (city is null || random.NextDouble() > 0.35)
         {
